% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_ct.R
\name{configure_admm}
\alias{configure_admm}
\title{ADMM algorithm configuration}
\usage{
configure_admm(rho = -1, tolerance = 1e-04, verbose = 0)
}
\arguments{
\item{rho}{Double. An ADMM parameter; coefficient of augmented term in the
Lagrangian function.}

\item{tolerance}{Double. Tolerance of ADMM convergence.}

\item{verbose}{Integer.}
}
\value{
a list of model parameters with class \code{admm_configuration}
}
\description{
ADMM algorithm configuration
}
