% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_ct.R
\name{interpolate_ct}
\alias{interpolate_ct}
\title{Interpolate (or extrapolate) ct estimates to intermediate design points}
\usage{
interpolate_ct(object, xout, ...)
}
\arguments{
\item{object}{A fitted object produced by \code{estimate_ct()} or
\code{cv_estimate_ct()}.}

\item{xout}{a vector of new positions at which Rt should be produced,
but where counts may not have been observed.}

\item{...}{additional arguments passed to methods.}
}
\value{
A vector or matrix of interpolated ct estimates.
}
\description{
Interpolate (or extrapolate) ct estimates to intermediate design points
}
