% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-cv.R
\name{plot.cv_smoothed_mult}
\alias{plot.cv_smoothed_mult}
\title{Plot cv_smoothed_mult}
\usage{
\method{plot}{cv_smoothed_mult}(x, which_lambda = c("cv_scores", "lambda.min", "lambda.1se"), ...)
}
\arguments{
\item{x}{result of cv_estimate_smooth of class \code{cv_smoothed_mult}}

\item{which_lambda}{select which ct's to plot.

If not provided, the cross validation score will be plotted. If provided a
list of lambda, the corresponding ct estimation will be plotted.

If provided a string, it
must be either one of \code{lambda.min}, \code{lambda.1se}, or \code{cv_scores}.
\itemize{
\item If provided \code{lambda.min}, plot ct which is generated from the lambda that
minimizes the cross validation score.
\item If provided \code{lambda.1se}, plot ct which is generated from the lambda whose
corresponding cross validation score is 1 standard error away of the
minimal cross validation score.
\item If provided \code{cv_scores}, plot the cross validation score.
\item If NULL, all estimated ct values are plotted.
}}

\item{...}{Not used.}
}
\value{
plot of cv scores
}
\description{
Plot cv_smoothed_mult
}
\examples{
y <- c(1, rnorm(100, dnorm(1:100, 50, 15) * 500 + 1))
cv <- cv_estimate_smooth(y, korder = 1, nfold = 3, nsol = 30)
plot(cv)
plot(cv, which_lambda = cv$lambda[1])
plot(cv, which_lambda = "lambda.min")
plot(cv, which_lambda = "lambda.1se")
plot(cv, NULL)
}
