% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{predict.smoothed_mult}
\alias{predict.smoothed_mult}
\title{Predict observed data using estimated ct}
\usage{
\method{predict}{smoothed_mult}(object, lambda = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{smoothed_mult} produced with \code{\link[=estimate_ct]{estimate_ct()}}.}

\item{lambda}{Select which lambdas from the object to use. If not provided
(the default), all are returned. Note that new lambdas not originally
used in the estimation procedure may be provided, but the results will be
calculated by linearly interpolating the estimated ct's.}

\item{...}{Not used.}
}
\value{
A vector or matrix of predicted case counts.
}
\description{
Given an object of class \code{smoothed_mult} produced with \code{\link[=estimate_ct]{estimate_ct()}},
calculate predicted observed cases for the estimated ct values.
Note: This function is not intended for "new x" or to produce forecasts, but
rather to examine how ct relates to observables.
}
\examples{
y <- c(1, rpois(100, dnorm(1:100, 50, 15)*500 + 1))
out <- estimate_ct(y, nsol = 10)
preds <- predict(out)
plot(y)
matlines(preds, lty = 1)
}
